Refactor build scripts to use shared manifest
Gradle tasks in multiple modules have been updated to utilize a shared manifest variable that is used across the tasks. This change reduces code duplication and makes it easier to manage and apply any changes to the manifest in the future. At the same time, the `unicorn-beans` build script has been migrated from Groovy to Kotlin DSL for better integration with Intellij and improved scripting capabilities.