Add documentation and refactor ConcurrencyUtils and UnicornAbstractThreadUtils
Added descriptive comments for the ConcurrencyUtils class in Java and attributes in UnicornAbstractThreadUtils that providing a brief summary of what they are used for. The specified documentation makes code easier to reuse and maintain for other developers. Also made a significant refactoring in QueryHelper of the Valentine project mainly focusing on improving the readability by rearranging arguments in several methods and removing unnecessary usage of the query object.