"Update Gradle scripts for Nexus publishing and versioning
The changes include renaming of OSSRH_USERNAME and OSSRH_PASSWORD to ossrhUsername and ossrhPassword in build.gradle.kts file to align with the preferred camel casing naming convention. This helps in maintaining a uniform code standard throughout the project.

In settings.gradle.kts, the version of springBoot and kotlin now fetches from gradle properties instead of fixed values. This improves the versatility of our project by allowing different builds to specify their respective dependencies' versions.

Lastly, the version name in com.lwohvye.java-conventions.gradle.kts has been updated to include '-Beta' indicating that the build is for development, not a release."