[#307] Update Redisson and Springdoc versions and remove Redisson module import
In this commit, the version of Redisson was updated from 3.23.4 to 3.25.0 and the version of Springdoc was updated from 2.2.0 to 2.3.0 in settings.gradle.kts file. The Redisson import was also removed from build.gradle.kts. This was done to keep the project up-to-date with the latest stable releases and ensure compatibility with other components. Additionally, the Redisson module was no longer necessary for the current project's needs, resulting in its removal.